reduce entity id size to 4 (u64->u32)

+++ src/entity_type.zig
@@ -1,5 +1,5 @@
 /// Entity ID type
-pub const EntityId = u64;
+pub const EntityId = u32;
 
 /// An opaque representation of a entity which can be used to interact with it
 pub const Entity = packed struct {
